(#20): Add new FrameController class(es).
To add flexibility, breaking out the time-driven aspect of the animator
by adding a new controller class that's passed to the animator that will
be used to make the animator perform its frame update process. We'll use this
facility soon to add a manual update controller that can by used to perform
frame updates at rates not directly correlated with wallclock time (e.g., when
wanting to write animation frames into a video file at a rate as fast as that
file can be written).